Separate file for view model.
